make gen_context.id an int
We print it with %d and this is more practical than changing every %d to
PRIu64something. It also makes sense to make it match check_context.id

+++ include/gen.h
@@ -67,7 +67,7 @@ struct gen_context {
 	struct type_store *store;
 	struct identifier *ns;
 
-	uint64_t id;
+	int id;
 
 	struct qbe_func *current;
 	const struct type *functype;
